+++ src/os_v5/process/process_errors.ts
@@ -1,0 +1,13 @@
+type ProcessErrorNotExecutable = {
+  readonly case: "not_executable"
+  readonly reason: string
+}
+type ProcessErrorTypes = ProcessErrorNotExecutable
+
+export class ProcessError extends Error {
+  public constructor(
+    public readonly error: ProcessErrorTypes,
+  ) {
+    super(error.reason)
+  }
+}
